Workaround for cross compile issue
This patch works around the issue with cross compiling kernels
on a different machine in which the tests are run from. If fixdep
binary is not present we know the pipeline workaround is in place
for cross compiled kernels, if so rebuild the external modules
on the target machine.

In addition, for upstream kernels we should pull in the latest fwts
version to include gcc10 fixes and add missing openssl-devel dependency
which is required to build efi_runtime.