perf requires libbpf
Perf needs libbpf.so, which is incorrectly provided by kernel-selftests-internal
package, so yum/dnf sometimes picks kernel-selftests-internal to satisfy the
dependencies. Since the given package *actually* does *not* provide the library,
linker can't find it, although it is OK at the rpm level. When linker can't find
the library, perf binary cannot be executed and thus all perf tests fail terribly.
As a workaround, add libbpf to the testcases explicitly.